re-named SPEMD -> PEMD, SPEMD_SMOOTH -> SPEMD warnings replaced with raise statements in default
